Move steer recovery planning into session harness
